Move NavBar component to a new NavBar directory
