+++ src/App.jsx
@@ -1,5 +1,5 @@
 import './App.css'
-import NavBar from './components/NavBar.jsx'
+import NavBar from './components/NavBar/NavBar.jsx'
 import ItemListContainer from './components/ItemListContainer.jsx'
 
 function App() {
